pyproject.toml
[tool.setuptools]
py-modules = []

[project]
name = "cryptotracker"
version = "0.1.0"
description = "CryptoTracker - веб-платформа криптоаналитики"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<5.0.0",
    "python-multipart>=0.0.6",
    "aiohttp>=3.9.0",
    "celery>=5.3.0",
    "websockets>=12.0",
    "python-telegram-bot>=20.7",
    "structlog>=23.2.0",
    "prometheus-client>=0.19.0",
    "aiosmtplib>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "email-validator>=2.1.0",

]



[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
